Added experimental driver, tried to optimize Jack's Rental
Added testrun.py to run from the command-line (instead of relying on a notebook).
Wrote a on-the-fly verssion of the Jack's Rental transitions, in an attempt to speed up, and refactored PolicyIteration to optionally take a functor as an arguments which computes expected rewards given (state, action).
The results don't match the previous outputs anymore though, have to investigate why.